+++ tests/test_mark.py
@@ -4,6 +4,8 @@ import textwrap
 import pytask
 import pytest
 from _pytask.mark import MarkGenerator
+from click.testing import CliRunner
+from pytask import cli
 from pytask import main
 
 
@@ -67,8 +69,9 @@ def test_ini_markers(tmp_path, config_name):
 
 @pytest.mark.end_to_end
 @pytest.mark.parametrize("config_name", ["pytask.ini", "tox.ini", "setup.cfg"])
-def test_markers_option(capsys, tmp_path, config_name):
-    tmp_path.joinpath(config_name).write_text(
+def test_markers_command(tmp_path, config_name):
+    config_path = tmp_path.joinpath(config_name)
+    config_path.write_text(
         textwrap.dedent(
             """
             [pytask]
@@ -80,13 +83,11 @@ def test_markers_option(capsys, tmp_path, config_name):
         )
     )
 
-    session = main({"paths": tmp_path, "markers": True})
+    runner = CliRunner()
+    result = runner.invoke(cli, ["markers", "-c", config_path.as_posix()])
 
-    assert session.exit_code == 0
-
-    captured = capsys.readouterr()
     for out in ["pytask.mark.a1", "pytask.mark.a2", "pytask.mark.nodescription"]:
-        assert out in captured.out
+        assert out in result.output
 
 
 @pytest.mark.end_to_end