+++ docs/changes.rst
@@ -11,6 +11,7 @@ all releases are available on `Anaconda.org <https://anaconda.org/pytask/pytask>
 
 - :gh:`16` reduces the traceback generated from tasks, failure section in report, fix
   error passing a file path to pytask, add demo to README.
+- :gh:`17` changes the interface to subcommands and adds ``"-c/--config"`` options.
 
 
 0.0.5 - 2020-08-12